+++ packages/frontend/src/pages/admin/other-settings.vue
@@ -36,6 +36,55 @@ SPDX-License-Identifier: AGPL-3.0-only
 						<template #caption>{{ i18n.ts.turnOffToImprovePerformance }}</template>
 					</MkSwitch>
 				</div>
+
+				<MkFolder :defaultOpen="true">
+					<template #icon><i class="ti ti-bolt"></i></template>
+					<template #label>Misskey® Fan-out Timeline Technology™ (FTT)</template>
+					<template v-if="enableFanoutTimeline" #suffix>Enabled</template>
+					<template v-else #suffix>Disabled</template>
+
+					<div class="_gaps_m">
+						<MkSwitch v-model="enableFanoutTimeline">
+							<template #label>{{ i18n.ts.enable }}</template>
+							<template #caption>{{ i18n.ts._serverSettings.fanoutTimelineDescription }}</template>
+						</MkSwitch>
+
+						<MkSwitch v-model="enableFanoutTimelineDbFallback">
+							<template #label>{{ i18n.ts._serverSettings.fanoutTimelineDbFallback }}</template>
+							<template #caption>{{ i18n.ts._serverSettings.fanoutTimelineDbFallbackDescription }}</template>
+						</MkSwitch>
+
+						<MkInput v-model="perLocalUserUserTimelineCacheMax" type="number">
+							<template #label>perLocalUserUserTimelineCacheMax</template>
+						</MkInput>
+
+						<MkInput v-model="perRemoteUserUserTimelineCacheMax" type="number">
+							<template #label>perRemoteUserUserTimelineCacheMax</template>
+						</MkInput>
+
+						<MkInput v-model="perUserHomeTimelineCacheMax" type="number">
+							<template #label>perUserHomeTimelineCacheMax</template>
+						</MkInput>
+
+						<MkInput v-model="perUserListTimelineCacheMax" type="number">
+							<template #label>perUserListTimelineCacheMax</template>
+						</MkInput>
+					</div>
+				</MkFolder>
+
+				<MkFolder :defaultOpen="true">
+					<template #icon><i class="ti ti-bolt"></i></template>
+					<template #label>Misskey® Reactions Buffering Technology™ (RBT)<span class="_beta">{{ i18n.ts.beta }}</span></template>
+					<template v-if="enableReactionsBuffering" #suffix>Enabled</template>
+					<template v-else #suffix>Disabled</template>
+
+					<div class="_gaps_m">
+						<MkSwitch v-model="enableReactionsBuffering">
+							<template #label>{{ i18n.ts.enable }}</template>
+							<template #caption>{{ i18n.ts._serverSettings.reactionsBufferingDescription }}</template>
+						</MkSwitch>
+					</div>
+				</MkFolder>
 			</div>
 		</FormSuspense>
 	</MkSpacer>
@@ -52,11 +101,20 @@ import { fetchInstance } from '@/instance.js';
 import { i18n } from '@/i18n.js';
 import { definePageMetadata } from '@/scripts/page-metadata.js';
 import MkSwitch from '@/components/MkSwitch.vue';
+import MkFolder from '@/components/MkFolder.vue';
+import MkInput from '@/components/MkInput.vue';
 
 const enableServerMachineStats = ref<boolean>(false);
 const enableIdenticonGeneration = ref<boolean>(false);
 const enableChartsForRemoteUser = ref<boolean>(false);
 const enableChartsForFederatedInstances = ref<boolean>(false);
+const enableFanoutTimeline = ref<boolean>(false);
+const enableFanoutTimelineDbFallback = ref<boolean>(false);
+const perLocalUserUserTimelineCacheMax = ref<number>(0);
+const perRemoteUserUserTimelineCacheMax = ref<number>(0);
+const perUserHomeTimelineCacheMax = ref<number>(0);
+const perUserListTimelineCacheMax = ref<number>(0);
+const enableReactionsBuffering = ref<boolean>(false);
 
 async function init() {
 	const meta = await misskeyApi('admin/meta');
@@ -64,6 +122,13 @@ async function init() {
 	enableIdenticonGeneration.value = meta.enableIdenticonGeneration;
 	enableChartsForRemoteUser.value = meta.enableChartsForRemoteUser;
 	enableChartsForFederatedInstances.value = meta.enableChartsForFederatedInstances;
+	enableFanoutTimeline.value = meta.enableFanoutTimeline;
+	enableFanoutTimelineDbFallback.value = meta.enableFanoutTimelineDbFallback;
+	perLocalUserUserTimelineCacheMax.value = meta.perLocalUserUserTimelineCacheMax;
+	perRemoteUserUserTimelineCacheMax.value = meta.perRemoteUserUserTimelineCacheMax;
+	perUserHomeTimelineCacheMax.value = meta.perUserHomeTimelineCacheMax;
+	perUserListTimelineCacheMax.value = meta.perUserListTimelineCacheMax;
+	enableReactionsBuffering.value = meta.enableReactionsBuffering;
 }
 
 function save() {
@@ -72,6 +137,13 @@ function save() {
 		enableIdenticonGeneration: enableIdenticonGeneration.value,
 		enableChartsForRemoteUser: enableChartsForRemoteUser.value,
 		enableChartsForFederatedInstances: enableChartsForFederatedInstances.value,
+		enableFanoutTimeline: enableFanoutTimeline.value,
+		enableFanoutTimelineDbFallback: enableFanoutTimelineDbFallback.value,
+		perLocalUserUserTimelineCacheMax: perLocalUserUserTimelineCacheMax.value,
+		perRemoteUserUserTimelineCacheMax: perRemoteUserUserTimelineCacheMax.value,
+		perUserHomeTimelineCacheMax: perUserHomeTimelineCacheMax.value,
+		perUserListTimelineCacheMax: perUserListTimelineCacheMax.value,
+		enableReactionsBuffering: enableReactionsBuffering.value,
 	}).then(() => {
 		fetchInstance(true);
 	});